API test for sportradar
to run first run npm i, then npm start. on a different terminal run node .\src\server.js. this method is a bit complicated, but it was the first way I got it to work. i will try to find a better way if we use this api. for now we know that this api works and have a working way of interacting with it.